+++ Src/FastMPH/Internals/HashCode3.cs
@@ -1,4 +1,4 @@
 namespace Genbox.FastMPH.Internals;
 
-internal delegate void HashCode3<in TKey>(TKey key, uint seed, Span<uint> hashes);
-internal delegate uint HashCode<in TKey>(TKey key, uint seed);
+internal delegate void HashCode3<in TKey>(TKey key, ulong seed, Span<uint> hashes);
+internal delegate ulong HashCode<in TKey>(TKey key, ulong seed);
